+++ assets/javascripts/kitten/components/form/form-actions.js
@@ -1,14 +1,29 @@
-import React from 'react'
+import React, { Component, PropTypes } from 'react'
 import classNames from 'classnames'
 
-export class FormActions extends React.Component {
+export class FormActions extends Component {
   render() {
-    const { className, children, ...others } = this.props
+    const { className, children, spreadOutAt, ...others } = this.props
+    const formActionsClassName = classNames(
+      'k-FormActions',
+      className,
+      {
+        [`k-FormActions--spreadOut@${spreadOutAt}`]: spreadOutAt
+      },
+    )
 
     return (
-      <div className={ classNames('k-FormActions', className) } { ...others }>
+      <div className={ formActionsClassName } { ...others }>
         { children }
       </div>
     )
   }
+}
+
+FormActions.propTypes = {
+  spreadOutAt: PropTypes.oneOf(['s-up'])
+}
+
+FormActions.defaultProps = {
+  spreadOutAt: 's-up'
 }