+++ assets/javascripts/kitten/karl/form/form-actions.js
@@ -5,23 +5,17 @@ import { Button } from 'kitten/components/buttons/button'
 export const KarlFormActions = props => {
   return (
     <FormActions { ...props }>
-      <Button className="karl-FormActions__button"
+      <Button className="karl-FormActions__button k-FormActions__item"
               tag="a"
               href="#"
               modifier="hydrogen">
         Button
       </Button>
 
-      <Button className="karl-FormActions__button"
+      <Button className="karl-FormActions__button k-FormActions__item"
               modifier="helium">
         Button
       </Button>
-
-      <Button className="karl-FormActions__button"
-              tag="input"
-              modifier="beryllium"
-              type="submit"
-              value="Button" />
     </FormActions>
   )
 }